feat: implement FUTEX_WAIT_BITSET and FUTEX_WAKE_BITSET with bitset matching for glibc pthreads compatibility (Item 341)
- Add bitset field to struct futex_waiter for bitset-aware futex operations
- Implement FUTEX_WAIT_BITSET: like FUTEX_WAIT but stores waiter's bitset mask
- Implement FUTEX_WAKE_BITSET: only wakes waiters whose bitset has non-empty
  intersection with the wake bitset
- Set FUTEX_WAIT waiter bitset to 0xFFFFFFFF (match-any semantics)
- Strip FUTEX_CLOCK_REALTIME flag along with FUTEX_PRIVATE_FLAG in op dispatch
- Add FUTEX_CLOCK_REALTIME, FUTEX_WAIT_BITSET_PRIVATE, FUTEX_WAKE_BITSET_PRIVATE
  constants
- Fix pre-existing -Werror warnings in cmd_ed.c and cmd_expr.c (format specifier
  type mismatches)

+++ src/shell/cmds/cmd_ed.c
@@ -166,7 +166,7 @@ void cmd_ed(const char *args) {
             case 'n':
             case 'N':
                 for (int i = addr1; i <= addr2 && i < ed_num_lines; i++)
-                    kprintf("%d\t%s\n", (unsigned long)(i+1), ed_buffer[i]);
+                    kprintf("%d\t%s\n", (int)(i+1), ed_buffer[i]);
                 ed_current = addr2;
                 break;
 
@@ -176,8 +176,8 @@ void cmd_ed(const char *args) {
                     const char *s = ed_buffer[i];
                     while (*s) {
                         if (*s == '\t') kprintf("\\t");
-                        else if (*s < 32) { kprintf("^%c", (unsigned long)(*s + 64)); }
-                        else kprintf("%c", (unsigned long)(uint8_t)*s);
+                        else if (*s < 32) { kprintf("^%c", (int)(*s + 64)); }
+                        else kprintf("%c", (int)(uint8_t)*s);
                         s++;
                     }
                     kprintf("$\n");
@@ -235,7 +235,7 @@ void cmd_ed(const char *args) {
                     }
                 }
                 if (replaced) ed_modified = 1;
-                kprintf("%d\n", (unsigned long)replaced);
+                kprintf("%d\n", (int)replaced);
                 break;
             }
 
@@ -252,7 +252,7 @@ void cmd_ed(const char *args) {
                     write_buf[wpos++] = '\n';
                 }
                 libc_vfs_write(ed_filename, write_buf, wpos);
-                kprintf("%d\n", (unsigned long)wpos);
+                kprintf("%d\n", (int)wpos);
                 ed_modified = 0;
                 break;
             }
